requirements.txt
graphviz==0.13.2
PyYAML==5.2
python-magic==0.4.15
pypandoc==1.4
colour==0.1.5
SecretColors==1.1.0
